\name{COA}
\alias{COA}
\title{construct a component orthogonal array with m components when m is prime or prime power}
\usage{COA(m)}
\arguments{
\item{m}{a positive integer, should be prime or prime power}
}
\value{a component orthogonal array with m components}
\description{construct a component orthogonal array with m components}
\examples{
COA(5)
}
