\name{fish}
\alias{fish}
\docType{data}
\title{
Data from an mixture order-of-addition experiment
}
\description{
Data from an mixture order-of-addition experiment
}
\usage{data(fish)}
\format{
  A data frame with 39 observations and following 7 variables.
  \describe{
    \item{\code{y}}{response variable}
    \item{\code{x1}}{first mixture component proportion}
    \item{\code{x2}}{second mixture component proportion}
	\item{\code{x3}}{third mixture component proportion}
	\item{\code{z12}}{first PWO variable}
	\item{\code{z13}}{second PWO variable}
	\item{\code{z23}}{third PWO variable}
  }
}
\examples{
data(fish)
}
\keyword{datasets}
