\name{oofa.scd}
\alias{oofa.scd}
\title{Order-of-addition Simplex Centroid Designs}
\usage{oofa.scd(m)}
\arguments{
\item{m}{number of components}
}
\value{An order-of-addition simplex centroid design}
\description{Construct an order-of-addition simplex centroid design with m components}
\examples{oofa.scd(4)}