\name{oofa.sld}
\alias{oofa.sld}
\title{Order-of-addition Simplex Lattice Designs}
\usage{
oofa.sld(m)
}
\arguments{
\item{m}{number of components}
}
\value{An order-of-addition simplex lattice design}
\description{Construct an order-of-addition simplex lattice design with m components}
\examples{
oofa.sld(4)
}
